Genomics experiment script
