+++ tasks/upload.py
@@ -6,7 +6,7 @@ from subprocess import call
 
 from invoke import task
 
-from tasks.util.config import get_faasm_config
+from tasks.util.endpoints import get_upload_host_port
 from tasks.util.env import FUNC_BUILD_DIR, PROJ_ROOT, RUNTIME_S3_BUCKET, FUNC_DIR, WASM_DIR, FAASM_SHARED_STORAGE_ROOT
 from tasks.util.genomics import INDEX_CHUNKS
 from tasks.util.upload_util import curl_file, upload_file_to_s3, upload_file_to_ibm
@@ -19,19 +19,6 @@ PYTHON_FUNC_DIR = join(FUNC_DIR, "python")
 def _get_s3_key(user, func):
     s3_key = "wasm/{}/{}/function.wasm".format(user, func)
     return s3_key
-
-
-def get_upload_host_port(host_in, port_in):
-    faasm_config = get_faasm_config()
-
-    if not host_in and faasm_config.has_section("Kubernetes"):
-        host = faasm_config["Kubernetes"].get("upload_host", "127.0.0.1")
-        port = faasm_config["Kubernetes"].get("upload_port", 8002)
-    else:
-        host = host_in if host_in else "127.0.0.1"
-        port = port_in if port_in else 8002
-
-    return host, port
 
 
 @task
